Create local network builder
