+++ integ_tester/pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "click>=8.3.1",
     "docker>=7.1.0",
     "pyre-extensions>=0.0.32",
+    "pyrefly>=0.53.0",
 ]
 
 [tool.pyrefly]
